compile on x86_64, and i assume other archs where sizeof returns something other than an unsigned long, which then causes problems with using kMin tested on x86 and x86_64
svn path=/trunk/kdebase/kioslave/; revision=352240

+++ sftp/kio_sftp.cpp
@@ -1470,7 +1470,7 @@ void sftpProtocol::slave_status() {
 bool sftpProtocol::getPacket(QByteArray& msg) {
 //    kdDebug(KIO_SFTP_DB) << "getPacket()" << endl;
     int len;
-    unsigned int msgLen;
+    size_t msgLen;
     char buf[4096];
 
     // Get the message length and type
